fixme/pyproject.toml
[tool.fastapi]
entrypoint = "server:fastapi_app"

[tool.poetry]
name = "FixMe"
version = "1.0.0"
description = "Doctor Octopus FIXME Server"
authors = ["Imran Sahil"]
readme = "readme.md"

[tool.poetry.dependencies]
python = ">=3.9,<4"
fastapi-slim = "0.128.8"
python-socketio = "5.13.0"
uvicorn = "0.35.0"
python-dotenv = "1.0.1"
aiofiles = "^25.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
server = "server:fastapi_app"
